#1860 Added nested inner classes and deleted HistoryDaoTestUtils
